Fix pinned tabs width when tab scrollbox is visible.
With this change, when the tab scrollbox is visible, selecting a pinned tab to the left is not very accurate, the mouse has to hover more to the left than needs to be. This is because the .tabbrowser-tab element is shifted to the left via the padding-left property.

I still need to figure out how to solve this, I solved this via pointer-events:none for .tab-background element, but I can't do that on the main .tabbrowser-tab element.